% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-fxns.R
\name{sensibo.pods}
\alias{sensibo.pods}
\title{Gets a list of the air conditioners (pods) IDs of the authenticated user.}
\usage{
sensibo.pods(key = getOption("sensibo.key"))
}
\arguments{
\item{key}{(character) API key from https://home.sensibo.com/me/api.}
}
\value{
A character vectors with available pods unique IDs.
}
\description{
Gets a list of the air conditioners (pods) IDs of the authenticated user.
}
\examples{
\dontrun{
# Assuming that a valid Sensibo Sky API Key was created on https://home.sensibo.com/me/api
# and added to a 'sensibo.sky' global option.
#
# options("sensibo.key" = <Your Sensibo API Key>)
 
## Getting the list of pods available to the user
pods.id <- sensibo.pods()

}
}
