% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibo.sky-package.R
\docType{package}
\name{sensibo.sky-package}
\alias{sensibo.sky-package}
\alias{sensibo.sky}
\title{Access to Sensibo Sky API V2 for Smart A/C Remote Control}
\description{
Access to Sensibo Sky API V2 for Smart A/C Remote Control
}
